+++ changelog.md
@@ -1,5 +1,7 @@
 # 0.3
 - [feat] new settings page
+- [feat] cluster dumps
+- [feat] move dumps via keys
 
 # 0.2
 - [feat] allow to scroll past the last line